+++ BlackStepper/DualStepperMotor.h
@@ -11,7 +11,7 @@ class DualStepperMotor {
 private:
 	BlackStepper leftStepper;
 	BlackStepper rightStepper;
-	int16_t turn_bias;
+	float turn_bias;
 
 public:
 	DualStepperMotor(
@@ -24,7 +24,7 @@ public:
 	void moveForward(uint64_t speed);
 	void moveBackward(uint64_t speed);
 	void run();
-	void setBias(int16_t bias);
+	void setBias(float bias);
 	void stop();
 
 	bool targetSpeedReached();
